refactor(api): refactor user entities (#9)
resolves #9
