+++ apps/api/src/app/users/entities/user.entity.ts
