+++ apps/api/src/app/users/users.service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class UsersService {
+  create() {
+    return 'This action adds a new user';
+  }
+
+  findAll() {
+    return `This action returns all users`;
+  }
+
+  findOne(id: number) {
+    return `This action returns a #${id} user`;
+  }
+
+  update(id: number) {
+    return `This action updates a #${id} user`;
+  }
+
+  remove(id: number) {
+    return `This action removes a #${id} user`;
+  }
+}
